+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/Components.java
@@ -23,6 +23,7 @@ import is.codion.swing.common.ui.component.button.ButtonBuilder;
 import is.codion.swing.common.ui.component.button.ButtonPanelBuilder;
 import is.codion.swing.common.ui.component.button.CheckBoxBuilder;
 import is.codion.swing.common.ui.component.button.CheckBoxMenuItemBuilder;
+import is.codion.swing.common.ui.component.button.ControlPanelBuilder;
 import is.codion.swing.common.ui.component.button.MenuBuilder;
 import is.codion.swing.common.ui.component.button.MenuItemBuilder;
 import is.codion.swing.common.ui.component.button.RadioButtonBuilder;
@@ -60,8 +61,10 @@ import is.codion.swing.common.ui.layout.Layouts;
 import javax.swing.JButton;
 import javax.swing.JComboBox;
 import javax.swing.JMenuItem;
+import javax.swing.JPanel;
 import javax.swing.JTextField;
 import javax.swing.JToggleButton;
+import javax.swing.JToolBar;
 import java.awt.BorderLayout;
 import java.math.BigDecimal;
 import java.nio.file.Path;
@@ -434,14 +437,14 @@ public final class Components {
 	/**
 	 * @return a {@link javax.swing.JToolBar} builder
 	 */
-	public static ToolBarBuilder toolBar() {
+	public static ControlPanelBuilder.ControlsStep<JToolBar, ToolBarBuilder> toolBar() {
 		return ToolBarBuilder.builder();
 	}
 
 	/**
 	 * @return a button panel builder
 	 */
-	public static ButtonPanelBuilder buttonPanel() {
+	public static ControlPanelBuilder.ControlsStep<JPanel, ButtonPanelBuilder> buttonPanel() {
 		return ButtonPanelBuilder.builder();
 	}